Fix aws profile override bug
Previously, when a session expired while using the `-nc` flags, any
future runs of the script would be unable to find a set of credentials
using the credentials file, since the `AWS_PROFILE` envvar was set to
the set of session credentials. This change caches the original
`AWS_PROFILE` value and specifically avoids using the session
credentials section of the credentials file.